Final Capstone Subsystem

+++ src/alert.cpp
@@ -9,7 +9,7 @@ bool send_alert(const std::string& message,
                 const std::string& severity,
                 const std::string& source,
                 const std::string& type,
-                int confidence,
+                double confidence,
                 const std::string& data_json) {
     const char* token = std::getenv("INTERNAL_ALERT_TOKEN");
     const char* url   = std::getenv("RTCT_INTERNAL_ALERT_URL");
